+++ Controllers/University/IUniversityController.cs
@@ -6,7 +6,6 @@ namespace Advisor.Controllers
     public interface IUniversityController
     {
         void HandlePreviousButtonClick();
-        void LoadData();
         void HandleFacultySelect(Faculty faculty);
         void HandleAddFaculty(object sender, EventArgs e);
         void HandleAddFacultyClick();
