+++ Controllers/University/UniversityController.cs
@@ -1,19 +1,42 @@
 ﻿using Advisor.Models;
 using System.Linq;
-using System.Collections.Generic;
 using System;
-using System.Collections.ObjectModel;
-using Advisor.Service.Statistics;
+using System.Web.Mvc;
+using System.Collections.Generic;
 
 namespace Advisor.Controllers
 {
-    public class UniversityController : IUniversityController
+    public class UniversityController : Controller, IUniversityController
     {
-        public University University { get; set; }
-
-        public UniversityController(University uni)
+        public UniversityController()
         {
-            University = uni;
+        }
+
+        [Route("universities/{id?}", Name = "universities_page")]
+        public ActionResult Index(int? id)
+        {
+            if (id != null)
+            {
+                University uni = DB.Instance.Universities.Where(u => u.Id == id).SingleOrDefault();
+                if (uni == null)
+                {
+                    return View("/Views/Shared/404.cshtml");
+                }
+
+                ViewBag.University = uni;
+
+                return View("/Views/University/University.cshtml");
+            }
+
+            return List();
+        }
+
+        private ActionResult List()
+        {
+            List<University> unis = DB.Instance.Universities.ToList();
+            ViewBag.Universities = unis;
+
+            return View("/Views/University/UniversityList.cshtml");
         }
 
         public void HandlePreviousButtonClick()
@@ -21,13 +44,6 @@ namespace Advisor.Controllers
             //UniversityView.Hide();
             //HomeView = new HomeView(new HomeController());
             //HomeView.ShowDialog();
-        }
-
-        public void LoadData()
-        {
-            LoadFacultyList();
-            LoadStats();
-            //UniversityView.University = University;
         }
 
         public void HandleFacultySelect(Faculty faculty)
@@ -77,21 +93,72 @@ namespace Advisor.Controllers
 
         private void LoadStats()
         {
-            StatsData statsData = new StatsData();
-            StatisticCalculator calculator = new StatisticCalculator();
-            List<Review> reviews = (from r in DB.Instance.Reviews
-                                           join p in DB.Instance.StudyPrograms on r.StudyProgram.Id equals p.Id
-                                           join f in DB.Instance.Faculties on p.Faculty.Id equals f.Id
-                                           join u in DB.Instance.Universities on f.University.Id equals u.Id
-                                    where u.Id == University.Id
-                                    select r).ToList();
-            statsData.AverageSalary = calculator.CalcReviewAverage(reviews, r => r.Salary, 2);
-            statsData.OveralRating = calculator.CalcReviewAverage(reviews, r => r.OveralRating, 1);
-            statsData.FacultyCount = University.Faculties.Count;
-            statsData.ReviewCount = reviews.Count;
+            //StatsData statsData = new StatsData();
+            //StatisticCalculator calculator = new StatisticCalculator();
+            //List<Review> reviews = (from r in DB.Instance.Reviews
+            //                               join p in DB.Instance.StudyPrograms on r.StudyProgram.Id equals p.Id
+            //                               join f in DB.Instance.Faculties on p.Faculty.Id equals f.Id
+            //                               join u in DB.Instance.Universities on f.University.Id equals u.Id
+            //                        where u.Id == University.Id
+            //                        select r).ToList();
+            //statsData.AverageSalary = calculator.CalcReviewAverage(reviews, r => r.Salary, 2);
+            //statsData.OveralRating = calculator.CalcReviewAverage(reviews, r => r.OveralRating, 1);
+            //statsData.FacultyCount = University.Faculties.Count;
+            //statsData.ReviewCount = reviews.Count;
 
             //UniversityView.StatsData = statsData;
         }
 
+        public ActionResult UniversityList()
+        {
+            List<University> unis = DB.Instance.Universities.ToList();
+            ViewBag.Universities = unis;
+
+            return View("/Views/University/UniversityList.cshtml");
+        }
+
+        public void HandleAddUniversityClick()
+        {
+            //AddFormView = new AddFormView();
+            //AddFormView.AddButtonClicked += HandleAddNewUniversity;
+            //AddFormView.TitleLabel.Text = "Add new university";
+            //AddFormView.ShowDialog();
+        }
+
+        public void HandleAddNewUniversity(object sender, EventArgs e)
+        {
+            //University uni = new University() {
+            //    Title = AddFormView.TitleInput.Text,
+            //    Description = AddFormView.DescriptionInput.Text
+            //};
+            //DB.Instance.Universities.Add(uni);
+            //DB.Instance.SaveChanges();
+            //HomeView.UniversityList.Items.Add(uni);
+            //AddFormView.Close();
+        }
+
+        public void HandleUniversitySelect(University uni)
+        {
+            //HomeView.Hide();
+            //UniversityView = new UniversityView(new UniversityController(uni));
+            //UniversityView.ShowDialog();
+        }
+
+        public void HandleSearchBoxChange()
+        {
+            //if (!string.IsNullOrEmpty(HomeView.SearchBox.Text))
+            //{
+            //    List<University> unis = DB.Instance.Universities.ToList();
+            //    HomeView.UniversityList.Items.Clear();
+            //    foreach (University uni in unis)
+            //    {
+            //        if (uni.Title.ToLower().Contains(HomeView.SearchBox.Text.ToLower()))
+            //        {
+            //            HomeView.UniversityList.Items.Add(uni); 
+            //        }
+            //    }
+            //}
+            //else LoadUniversityList();
+        }
     }
 }
