+++ Views/Shared/404.cshtml
@@ -1,0 +1,3 @@
+﻿<div>
+    <h3>Oops! The page you requested does not exist.</h3>
+</div>
