+++ Views/Shared/_Layout.cshtml
@@ -22,7 +22,7 @@
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav navbar-right">
-                    <li>@Html.ActionLink("Universities", "UniversityList", "UniversityList")</li>
+                    <li><a href="@Url.RouteUrl("universities_page")">Universties</a></li>
                 </ul>
                 @Html.Partial("_LoginPartial")
             </div>
