+++ Views/University/University.cshtml
@@ -1,0 +1,7 @@
+﻿@using Advisor.Models
+@{
+    ViewBag.Title = "Universities";
+}
+<div class="w-100 mt-100">
+    <h2 class="mb-20">@ViewBag.University.Title</h2>
+</div>
