--- Views/UniversityList/UniversityList.cshtml
@@ -1,34 +1,0 @@
-﻿@using Advisor.Models
-@{
-    ViewBag.Title = "Universities";
-}
-<div class="w-100 mt-100">
-    <h2 class="mb-20">Which university are you interested in at the moment?</h2>
-    <div class="d-flex flex-wrap mt-50">
-        @foreach (University uni in ViewBag.Universities)
-        {
-        <div class="university-container card">
-            <a href="#">
-                <h3>@uni.Title</h3>
-            </a>
-            <div class="d-flex">
-                <div class="d-flex mr-5">
-                    <p class="mr-5">
-                        <i class="fas fa-globe"></i>
-                    </p>
-                    <p class="mr-5">
-                        <a target="_blank" href="@uni.Website">@uni.Website</a>
-                    </p>
-                </div>
-                <div class="d-flex">
-                    <p class="mr-5">
-                        <i class="fas fa-graduation-cap"></i>
-                    </p>
-                    <p>@uni.Faculties.Count faculties</p>
-                </div>
-            </div>
-            <p>@uni.Description</p>
-        </div>
-        }
-    </div>
-</div>
